# openapi/mopl-api.yaml
openapi: "3.1.0"
info:
  title: "모두의 플리 (Mopl) API"
  description: "프로토타입 OpenAPI를 기준으로 팀이 합의한 요청·응답 규칙을 반영한 API 계약입니다."
  version: "1.0"
servers:
  -
    url: "/"
    description: "현재 서버"
  -
    url: "http://localhost:8080"
    description: "로컬 서버"
security:
  -
    BearerAuth: []
...
paths:
...
  "/api/follows":
    post:
      tags:
        - "팔로우 관리"
      summary: "팔로우"
      operationId: "createFollow"
      security:
        -
          BearerAuth: []
          CsrfToken: []
      requestBody:
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/FollowRequest"
        required: true
      responses:
        "201":
          description: "새 팔로우 관계 생성"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/FollowDto"
        "200":
          description: "이미 팔로우 중이면 기존 관계 반환"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/FollowDto"
        "400":
          description: "잘못된 요청"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "401":
          description: "인증 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          description: "서버 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
  "/api/follows/followed-by-me":
    get:
      tags:
        - "팔로우 관리"
      summary: "특정 유저를 내가 팔로우하는지 여부 조회"
      description: "팔로우 중이면 FollowDto(id 포함)를 반환합니다. 팔로우하지 않은 경우 404를 반환합니다."
      operationId: "isFollowedByMe"
      parameters:
        -
          name: "followeeId"
          in: "query"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "성공"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/FollowDto"
        "400":
          description: "잘못된 요청"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "401":
          description: "인증 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "404":
          description: "해당 리소스 없음"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          description: "서버 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/api/follows/count":
    get:
      tags:
        - "팔로우 관리"
      summary: "특정 유저의 팔로워 수 조회"
      operationId: "getFollowerCount"
      parameters:
        -
          name: "followeeId"
          in: "query"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "성공"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/FollowerCountResponse"
        "400":
          description: "잘못된 요청"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "401":
          description: "인증 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          description: "서버 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/api/follows/followers":
    get:
      tags:
        - "팔로우 관리"
      summary: "특정 유저의 팔로워 목록 조회 (커서 페이지네이션)"
      description: "지정한 followeeId 를 팔로우한 사용자들의 목록을 최근순(created_at DESC)으로 반환합니다."
      operationId: "getFollowers"
      security:
        -
          BearerAuth: []
      parameters:
        -
          name: "followeeId"
          in: "query"
          required: true
          schema:
            type: "string"
            format: "uuid"
        -
          name: "cursor"
          in: "query"
          description: "다음 페이지 커서 (마지막 followedAt 을 Base64 로 인코딩). idAfter 와 함께 제공해야 합니다."
          required: false
          schema:
            type: "string"
        -
          name: "idAfter"
          in: "query"
          description: "같은 followedAt 값 내 UUID 타이브레이커. cursor 와 함께 제공해야 합니다."
          required: false
          schema:
            type: "string"
            format: "uuid"
        -
          name: "limit"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int32"
            minimum: 1
            maximum: 100
        -
          name: "sortBy"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "followedAt"
            enum:
              - "followedAt"
        -
          name: "sortDirection"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "DESCENDING"
            description: |
              현재 팔로우 목록 조회는 최신순(DESCENDING)만 지원한다.
              오래된 순 요청(ASCENDING)은 400 을 반환한다.
            enum:
              - "DESCENDING"
      responses:
        "200":
          description: "성공"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CursorResponseFollowUserItemDto"
        "400":
          description: "잘못된 요청 (파라미터 검증 실패, 잘못된 커서, cursor/idAfter 부정합)"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "401":
          description: "인증 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          description: "서버 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/api/follows/followings":
    get:
      tags:
        - "팔로우 관리"
      summary: "특정 유저의 팔로잉 목록 조회 (커서 페이지네이션)"
      description: "지정한 followerId 가 팔로우 중인 사용자들의 목록을 최근순(created_at DESC)으로 반환합니다."
      operationId: "getFollowings"
      security:
        -
          BearerAuth: []
      parameters:
        -
          name: "followerId"
          in: "query"
          required: true
          schema:
            type: "string"
            format: "uuid"
        -
          name: "cursor"
          in: "query"
          description: "다음 페이지 커서. idAfter 와 함께 제공해야 합니다."
          required: false
          schema:
            type: "string"
        -
          name: "idAfter"
          in: "query"
          description: "같은 followedAt 값 내 UUID 타이브레이커. cursor 와 함께 제공해야 합니다."
          required: false
          schema:
            type: "string"
            format: "uuid"
        -
          name: "limit"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int32"
            minimum: 1
            maximum: 100
        -
          name: "sortBy"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "followedAt"
            enum:
              - "followedAt"
        -
          name: "sortDirection"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "DESCENDING"
            description: |
              현재 팔로우 목록 조회는 최신순(DESCENDING)만 지원한다.
              오래된 순 요청(ASCENDING)은 400 을 반환한다.
            enum:
              - "DESCENDING"
      responses:
        "200":
          description: "성공"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CursorResponseFollowUserItemDto"
        "400":
          description: "잘못된 요청 (파라미터 검증 실패, 잘못된 커서, cursor/idAfter 부정합)"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "401":
          description: "인증 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          description: "서버 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/api/follows/recommendations":
    get:
      tags:
        - "팔로우 관리"
      summary: "팔로우 추천 목록 조회 (친구의 친구 기반)"
      description: "요청자가 팔로우 중인 사용자들이 팔로우하는 대상을 후보로, 공통 팔로잉 수(commonFollowingCount) DESC → userId DESC 정렬로 반환합니다. 자기 자신과 이미 팔로우한 사용자는 제외합니다."
      operationId: "getFollowRecommendations"
      security:
        -
          BearerAuth: []
      parameters:
        -
          name: "cursor"
          in: "query"
          description: "다음 페이지 커서 (마지막 commonFollowingCount 를 Base64 로 인코딩). idAfter 와 함께 제공해야 합니다."
          required: false
          schema:
            type: "string"
        -
          name: "idAfter"
          in: "query"
          description: "같은 commonFollowingCount 값 내 userId 타이브레이커. cursor 와 함께 제공해야 합니다."
          required: false
          schema:
            type: "string"
            format: "uuid"
        -
          name: "limit"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int32"
            minimum: 1
            maximum: 100
        -
          name: "sortBy"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "commonFollowingCount"
            enum:
              - "commonFollowingCount"
        -
          name: "sortDirection"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "DESCENDING"
            enum:
              - "DESCENDING"
      responses:
        "200":
          description: "성공"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CursorResponseFollowRecommendationItemDto"
        "400":
          description: "잘못된 요청 (파라미터 검증 실패, 잘못된 커서, cursor/idAfter 부정합)"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "401":
          description: "인증 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          description: "서버 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
  "/api/follows/{followId}":
    delete:
      tags:
        - "팔로우 관리"
      summary: "팔로우 취소"
      description: "API 요청자 본인의 팔로우만 취소할 수 있습니다."
      operationId: "cancelFollow"
      security:
        -
          BearerAuth: []
          CsrfToken: []
      parameters:
        -
          name: "followId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "204":
          description: "성공"
        "400":
          description: "잘못된 요청"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "401":
          description: "인증 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "403":
          description: "권한 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        "500":
          description: "서버 오류"
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: "object"
      properties:
        exceptionName:
          type: "string"
          description: "예외 이름"
        message:
          type: "string"
          description: "오류 메시지"
        details:
          type: "object"
          additionalProperties:
            type: "string"
          description: "오류 부가 정보"
        errorCode:
          type: "string"
          description: "애플리케이션 오류 코드"
          example: "COMMON_400_1"
      required:
        - "exceptionName"
        - "errorCode"
        - "message"
        - "details"
...
    UserSummary:
      type: "object"
      description: |
        여러 도메인의 응답에서 공통으로 사용하는 사용자 요약.
        name / profileImageUrl 은 User 도메인 연동이 완료되기 전까지 null 로 반환될 수 있어
        OpenAPI 3.1 표준 nullable 표기(["string", "null"])를 사용한다.
      properties:
        userId:
          type: "string"
          format: "uuid"
          description: "사용자 ID"
        name:
          type:
            - "string"
            - "null"
          description: "사용자 이름 (User 도메인 연동 전까지 null 가능)"
        profileImageUrl:
          type:
            - "string"
            - "null"
          description: "사용자 프로필 이미지 URL (User 도메인 연동 전까지 null 가능)"
      required:
        - "userId"
...
    FollowRequest:
      type: "object"
      properties:
        followeeId:
          type: "string"
          format: "uuid"
          description: "팔로우 대상 사용자 ID"
      required:
        - "followeeId"
    FollowDto:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "팔로우 ID"
        followeeId:
          type: "string"
          format: "uuid"
          description: "팔로우 대상 사용자 ID"
        followerId:
          type: "string"
          format: "uuid"
          description: "팔로워 사용자 ID"
      required:
        - "followeeId"
        - "followerId"
        - "id"
    FollowUserItemDto:
      type: "object"
      description: "팔로워/팔로잉 목록의 아이템. followers 응답에서 user 는 팔로워, followings 응답에서 user 는 팔로우 대상."
      properties:
        followId:
          type: "string"
          format: "uuid"
          description: "팔로우 관계 ID (unfollow 시 사용)"
        user:
          "$ref": "#/components/schemas/UserSummary"
          description: "상대 사용자 요약. name/profileImageUrl 은 User 도메인 연동 전까지 null."
        followedAt:
          type: "string"
          format: "date-time"
          description: "팔로우 관계 생성 시각"
      required:
        - "followId"
        - "user"
        - "followedAt"
    FollowRecommendationItemDto:
      type: "object"
      description: "팔로우 추천 목록의 아이템. user 는 요청자에게 추천되는 사용자, commonFollowingCount 는 요청자와 겹치는 매개자 수."
      properties:
        user:
          "$ref": "#/components/schemas/UserSummary"
          description: "추천 대상 사용자 요약. 대응하는 사용자가 없으면 UNKNOWN fallback."
        commonFollowingCount:
          type: "integer"
          format: "int64"
          description: "요청자와 이 후보가 공통으로 팔로우하는 사용자 수. 클수록 취향 겹칠 가능성 높음."
      required:
        - "user"
        - "commonFollowingCount"
...
    CursorResponseFollowRecommendationItemDto:
      type: "object"
      properties:
        data:
          type: "array"
          description: "팔로우 추천 목록"
          items:
            "$ref": "#/components/schemas/FollowRecommendationItemDto"
        nextCursor:
          type:
            - "string"
            - "null"
          description: "다음 커서 (마지막 commonFollowingCount 를 Base64 로 인코딩)"
        nextIdAfter:
          type:
            - "string"
            - "null"
          format: "uuid"
          description: "다음 요청의 userId 타이브레이커"
        hasNext:
          type: "boolean"
          description: "다음 데이터가 있는지 여부"
        totalCount:
          type: "integer"
          format: "int64"
          description: "현재 페이지 데이터 개수. 추천은 무한 스크롤 UI 라 정확 집계 대신 페이지 크기로 대체."
        sortBy:
          type: "string"
          description: "정렬 기준"
        sortDirection:
          type: "string"
          description: "정렬 방향"
          enum:
            - "DESCENDING"
      required:
        - "data"
        - "hasNext"
        - "sortBy"
        - "sortDirection"
        - "totalCount"
    CursorResponseFollowUserItemDto:
      type: "object"
      properties:
        data:
          type: "array"
          description: "팔로워/팔로잉 목록"
          items:
            "$ref": "#/components/schemas/FollowUserItemDto"
        nextCursor:
          type:
            - "string"
            - "null"
          description: "다음 커서"
        nextIdAfter:
          type:
            - "string"
            - "null"
          format: "uuid"
          description: "다음 요청의 보조 커서"
        hasNext:
          type: "boolean"
          description: "다음 데이터가 있는지 여부"
        totalCount:
          type: "integer"
          format: "int64"
          description: "총 데이터 개수"
        sortBy:
          type: "string"
          description: "정렬 기준"
        sortDirection:
          type: "string"
          description: "정렬 방향"
          enum:
            - "ASCENDING"
            - "DESCENDING"
      required:
        - "data"
        - "hasNext"
        - "sortBy"
        - "sortDirection"
        - "totalCount"
...
    FollowerCountResponse:
      type: "object"
      properties:
        count:
          type: "integer"
          format: "int64"
          minimum: 0
          description: "팔로워 수"
      required:
        - "count"
  securitySchemes:
    BearerAuth:
      type: "http"
      description: "로그인 후 발급받은 JWT 액세스 토큰"
      scheme: "bearer"
      bearerFormat: "JWT"
    CsrfToken:
      type: "apiKey"
      description: "GET /api/auth/csrf-token 호출 후 XSRF-TOKEN 쿠키로 받은 값을 전달합니다."
      name: "X-XSRF-TOKEN"
      in: "header"